pyproject.toml
[tool.poetry]
name = "enterprise-ai-copilot"
version = "0.1.0"
description = "Enterprise AI Knowledge Copilot using FastAPI + Databricks Mosaic AI"
authors = ["ZoneOne <dev@zoneone.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.116.0,<1.0.0"
uvicorn = {extras = ["standard"], version = ">=0.31.0,<1.0.0"}
pydantic = ">=2.9.0,<3.0.0"
pydantic-settings = ">=2.5.0,<3.0.0"
mlflow = ">=3.0.0,<4.0.0"
httpx = ">=0.27.0,<1.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
psycopg = {extras = ["binary"], version = ">=3.2.1,<4.0.0"}
requests = ">=2.32.0,<3.0.0"
PyJWT = ">=2.9.0,<3.0.0"
databricks-sdk = ">=0.50.0,<1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = ">=4.0.0,<5.0.0"
ruff = ">=0.11.0,<1.0.0"
pytest = ">=8.3.0,<9.0.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
